Table Model update
Added Board to the database and established relationships with Group. Additionally, moved db_setup.py to extensions.
