+++ .idea/hindsite.iml
@@ -3,4 +3,8 @@
   <component name="NewModuleRootManager">
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
+  <component name="PyDocumentationSettings">
+    <option name="format" value="PLAIN" />
+    <option name="myDocStringFormat" value="Plain" />
+  </component>
 </module>
