+++ .idea/workspace.xml
@@ -5,21 +5,16 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="257b68f9-7575-40dc-9d1c-f6cbe7a3603c" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/.gitignore" beforeDir="false" afterPath="$PROJECT_DIR$/.gitignore" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/hindsite.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/hindsite.iml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/hindsite/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/__init__.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/hindsite/authenticate.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/authenticate.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/hindsite/db_setup.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/db_setup.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/hindsite/db_setup.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/hindsite/extensions.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/extensions.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/hindsite/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/main.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/hindsite/routes.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/routes.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/hindsite/sql_query.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/sql_query.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/hindsite/sql_update.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/sql_update.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/hindsite/tables/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/tables/__init__.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/hindsite/tables/group.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/tables/group.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/hindsite/tables/password.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/tables/password.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/hindsite/tables/user.py" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/tables/user.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/hindsite/templates/base.html" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/templates/base.html" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/hindsite/templates/index.html" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/templates/index.html" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/hindsite/templates/sign-in.html" beforeDir="false" afterPath="$PROJECT_DIR$/hindsite/templates/sign-in.html" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -161,7 +156,7 @@
       <breakpoints>
         <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
           <url>file://$PROJECT_DIR$/hindsite/main.py</url>
-          <line>11</line>
+          <line>13</line>
           <option name="timeStamp" value="1" />
         </line-breakpoint>
         <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
@@ -178,11 +173,6 @@
           <url>file://$PROJECT_DIR$/hindsite/main.py</url>
           <line>4</line>
           <option name="timeStamp" value="5" />
-        </line-breakpoint>
-        <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
-          <url>file://$PROJECT_DIR$/hindsite/tables/user.py</url>
-          <line>3</line>
-          <option name="timeStamp" value="6" />
         </line-breakpoint>
       </breakpoints>
     </breakpoint-manager>
