+++ hindsite/main.py
@@ -5,8 +5,9 @@ Activation point for the application.
 import os
 
 from flask_bootstrap import Bootstrap5
-from hindsite import app
+from hindsite import create_app
 
+app = create_app()
 
 bootstrap = Bootstrap5(app)
