+++ hindsite/tables/__init__.py
@@ -5,3 +5,4 @@ Package specification for tables, importing all the primary tables.
 from .user import User
 from .password import Password
 from .group import Group
+from .board import Board
